+++ src/lineup/ALineUpView2.ts
@@ -1,0 +1,601 @@
+/**
+ * Created by Samuel Gratzl on 29.01.2016.
+ */
+import {AView, EViewMode, IViewContext, ISelection, ViewWrapper} from '../View';
+import LineUp, {ILineUpConfig} from 'lineupjs/src/lineup';
+import {deriveColors} from 'lineupjs/src/';
+import {ScaleMappingFunction, createSelectionDesc} from 'lineupjs/src/model';
+import ValueColumn from 'lineupjs/src/model/ValueColumn';
+import NumberColumn from 'lineupjs/src/model/NumberColumn';
+import {IBoxPlotData} from 'lineupjs/src/model/BoxPlotColumn';
+import {LocalDataProvider, } from 'lineupjs/src/provider';
+import * as d3 from 'd3';
+import {resolve, IDType} from 'phovea_core/src/idtype';
+import {list as rlist, RangeLike, Range} from 'phovea_core/src/range';
+import * as cmds from './cmds';
+import {saveNamedSet} from '../storage';
+import {showErrorModalDialog} from '../Dialogs';
+import {LineUpRankingButtons} from './LineUpRankingButtons';
+import {LineUpSelectionHelper, array_diff} from './LineUpSelectionHelper';
+import IScore, {IScoreRow} from './IScore';
+import {stringCol} from './desc';
+
+
+export function useDefaultLayout(instance: LineUp) {
+  instance.data.deriveDefault();
+  //insert selection column
+  instance.data.insert(instance.data.getRankings()[0], 1, createSelectionDesc());
+}
+
+
+export abstract class ALineUpView2 extends AView {
+
+  resolver: (d: any) => void;
+
+  protected lineup: LineUp;
+
+  protected additionalScoreParameter: any = null;
+
+  private initLineUpPromise: Promise<any>;
+
+  private config : ILineUpConfig = {
+    renderingOptions: {
+      histograms: true
+    },
+    header: {
+      rankingButtons: ($node: d3.Selection<any>) => {
+        const rb = new LineUpRankingButtons(this.lineup, $node, this.idType, this.additionalScoreParameter);
+        rb.on(LineUpRankingButtons.SAVE_NAMED_SET, (event, order, name, description) => {
+          this.saveNamedSet(order, name, description);
+        });
+        rb.on(LineUpRankingButtons.ADD_SCORE_COLUMN, (event, scoreImpl, scorePlugin) => {
+          this.addScoreColumn(scoreImpl, scorePlugin);
+        });
+        return rb;
+      }
+    },
+    body: {}
+  };
+
+  protected idType: IDType;
+
+  /**
+   * Stores the ranking data when collapsing columns on modeChange()
+   * @type {any}
+   */
+  private dump: any = null;
+
+  /**
+   * DOM element with message when no data is available
+   */
+  protected $nodata: d3.Selection<any>;
+
+  /**
+   * DOM element for LineUp stats in parameter UI
+   */
+  private $params: d3.Selection<ViewWrapper>;
+
+  private selectionHelper: LineUpSelectionHelper;
+
+  protected idAccessor = (d) => d._id;
+
+  private scoreAccessor = (row: any, index: number, id: string, desc: any) => {
+    const rowId = this.idAccessor(row);
+    let r = (desc.scores && typeof desc.scores[rowId] !== 'undefined') ? desc.scores[rowId] : (typeof desc.missingValue !== 'undefined' ? desc.missingValue : null);
+    if (desc.type === 'categorical') {
+      r = String(r); //even null values
+    }
+    return r;
+  }
+
+  constructor(context: IViewContext, protected selection: ISelection, parent: Element, private options?: {}) {
+    super(context, parent, options);
+
+    this.$node.classed('lineup', true);
+
+    this.context.ref.value.data = new Promise((resolve) => {
+      this.resolver = resolve;
+    });
+    //context.graph.findOrAddObject()
+
+    this.$nodata = this.$node.append('p')
+      .classed('nodata', true)
+      .classed('hidden', true)
+      .text('No data found for selection and parameter.');
+  }
+
+  init() {
+    super.init();
+    //this.build(); // will be called from update()
+    this.update();
+  }
+
+  private initSelectionHelper() {
+    this.selectionHelper = new LineUpSelectionHelper(this.lineup, this.idType, this.idAccessor);
+    this.selectionHelper.on(LineUpSelectionHelper.SET_ITEM_SELECTION, (event, selection) => {
+      this.setItemSelection(selection);
+    });
+    this.selectionHelper.init();
+    this.updateLineUpStats();
+  }
+
+  changeSelection(selection: ISelection) {
+    super.changeSelection(selection);
+    this.selection = selection;
+    this.handleSelectionColumns(this.selection);
+  }
+
+  setItemSelection(selection: ISelection) {
+    this.selectionHelper.setItemSelection(selection);
+    this.updateLineUpStats();
+    super.setItemSelection(selection);
+  }
+
+  buildParameterUI($parent: d3.Selection<any>, onChange: (name: string, value: any)=>Promise<any>) {
+    super.buildParameterUI($parent, onChange);
+
+    // used for LineUp stats
+    this.$params = $parent.append('div')
+      .classed('form-group', true)
+      .append('p');
+  }
+
+  getParameter(name: string): any {
+    return super.getParameter(name);
+  }
+
+  setParameter(name: string, value: any) {
+    return super.setParameter(name, value);
+  }
+
+  /**
+   * Expand/collapse certain columns on mode change.
+   * Expand = focus view
+   * Collapse = context view
+   * @param mode
+   */
+  modeChanged(mode: EViewMode) {
+    super.modeChanged(mode);
+    if (!this.lineup) {
+      return;
+    }
+
+    // collapse all columns
+    const data = this.lineup.data;
+
+    if (mode === EViewMode.FOCUS) {
+      if (this.dump) {
+        const r = data.getRankings()[0];
+        r.children.forEach((c) => {
+          if (c.id in this.dump) {
+            c.setWidth(this.dump[c.id]);
+          }
+        });
+      }
+      this.dump = null;
+
+    } else if (this.dump === null) {
+      const r = data.getRankings()[0];
+      const s = r.getSortCriteria();
+      const labelColumn = r.children.filter((c) => c.desc.type === 'string')[0];
+
+      this.dump = {};
+      r.children.forEach((c) => {
+        if (c === labelColumn ||
+          c === s.col ||
+          c.desc.type === 'rank' ||
+          c.desc.type === 'selection' ||
+          (<any>c.desc).column === 'id' // = Ensembl column
+        ) {
+          // keep these columns
+        } else {
+          this.dump[c.id] = c.getWidth();
+          c.setWidth(0);
+        }
+      });
+    }
+  }
+
+  /**
+   * Get sub type for named sets
+   * @returns {{key: string, value: string}}
+   */
+  protected getSubType() {
+    return {
+      key: '',
+      value: ''
+    };
+  }
+
+  private saveNamedSet(order, name, description) {
+    const r = this.selectionHelper.rows;
+    const ids = rlist(order.map((i) => this.idAccessor(r[i])).sort(d3.ascending));
+    saveNamedSet(name, this.idType, ids, this.getSubType(), description).then((d) => {
+      console.log('saved', d);
+      this.fire(AView.EVENT_UPDATE_ENTRY_POINT, this.idType, d);
+    });
+  }
+
+  protected handleSelectionColumns(selection) {
+    this.initLineUpPromise.then(() => {
+      this.handleSelectionColumnsImpl(selection);
+    });
+  }
+
+  protected handleSelectionColumnsImpl(selection) {
+    const selectedIds = selection.range.dim(0).asList();
+
+    const ranking = this.lineup.data.getLastRanking();
+    const usedCols = ranking.flatColumns.filter((d) => (<any>d.desc).selectedId !== -1);
+    const lineupColIds = usedCols
+      .map((d) => (<any>d.desc).selectedId)
+      .filter((d) => d !== undefined);
+
+    // compute the difference
+    const diffAdded = array_diff(selectedIds, lineupColIds);
+    const diffRemoved = array_diff(lineupColIds, selectedIds);
+
+    // add new columns to the end
+    if (diffAdded.length > 0) {
+      //console.log('add columns', diffAdded);
+      diffAdded.forEach((id) => {
+        this.getSelectionColumnDesc(id)
+          .then((columnDesc) => {
+            this.withoutTracking(() => {
+              this.addColumn(columnDesc, this.loadSelectionColumnData.bind(this), id, true); // true == withoutTracking
+            });
+          });
+      });
+    }
+
+    // remove deselected columns
+    if (diffRemoved.length > 0) {
+      this.withoutTracking(() => {
+        //console.log('remove columns', diffRemoved);
+        diffRemoved.forEach((id) => {
+          const col = usedCols.filter((d) => (<any>d.desc).selectedId === id)[0];
+          ranking.remove(col);
+        });
+      });
+    }
+  }
+
+  protected getSelectionColumnId(id) {
+    // hook
+    return `col_${id}`;
+  }
+
+  protected getSelectionColumnLabel(id) {
+    // hook
+    return new Promise((resolve) => {
+      const label = `Selection ${id}`;
+      resolve(label);
+    });
+  }
+
+  protected getSelectionColumnDesc(id) {
+    return this.getSelectionColumnLabel(id)
+      .then((label: string) => {
+        return stringCol(this.getSelectionColumnId(id), label, true, 50, id);
+      });
+  }
+
+  protected addColumn(colDesc, loadColumnData: (id) => Promise<IScoreRow<any>[]>, id = -1, withoutTracking = false) {
+    const ranking = this.lineup.data.getLastRanking();
+    const colors = this.getAvailableColumnColors(ranking);
+
+    colDesc.color = colors.shift(); // get and remove color from list
+    colDesc.accessor = this.scoreAccessor;
+    const provider = <LocalDataProvider>this.lineup.data;
+    provider.pushDesc(colDesc);
+    const col = <ValueColumn<any>>this.lineup.data.push(ranking, colDesc);
+
+    const loadPromise = loadColumnData(id);
+    // error handling
+    loadPromise
+      .catch(showErrorModalDialog)
+      .catch(() => {
+        ranking.remove(col);
+      });
+
+    // success
+    loadPromise
+    // map selection rows
+      .then((rows: IScoreRow<any>[]) => {
+        if (id !== -1) {
+          return this.mapSelectionRows(rows);
+        }
+        return rows;
+      })
+      // convert to score array to object to use in LineUp
+      .then((rows: IScoreRow<any>[]) => {
+        const r: { [id: string]: number } = {};
+
+
+
+
+        rows.forEach((row) => {
+
+          r[this.selectionHelper.underscoreIdAccessor(row.id)] = row.score;
+        });
+        return r;
+      })
+      .then((scores: { [id: string]: any }) => {
+        colDesc.scores = scores;
+
+        if (colDesc.type === 'number') {
+          const ncol = <NumberColumn>col;
+          if (!(colDesc.constantDomain)) { //create a dynamic range if not fixed
+            colDesc.domain = d3.extent(<number[]>(d3.values(scores)));
+          }
+          // add selection columns without tracking changes
+          if (withoutTracking) {
+            this.withoutTracking(() => {
+              ncol.setMapping(new ScaleMappingFunction(colDesc.domain));
+            });
+            // however, track changes in score columns
+          } else {
+            ncol.setMapping(new ScaleMappingFunction(colDesc.domain));
+          }
+        } else if (colDesc.type === 'boxplot') {
+          const values = <IBoxPlotData[]>d3.values(scores);
+          //HACK we know that the domain of the description is just referenced, so we can update it by changing values!
+          if (!(colDesc.constantDomain)) { //create a dynamic range if not fixed
+            colDesc.domain[0] = d3.min(values, (d) => d.min);
+            colDesc.domain[1] = d3.max(values, (d) => d.max);
+          }
+        }
+        col.setLoaded(true);
+        this.lineup.update();
+      });
+
+    return col;
+  }
+
+  protected mapSelectionRows(rows: IScoreRow<any>[]) {
+    // hook
+    return rows;
+  }
+
+  protected getAvailableColumnColors(ranking = this.lineup.data.getLastRanking()) {
+    const colors = d3.scale.category10().range().slice();
+    // remove colors that are already in use from the list
+    ranking.flatColumns.forEach((d) => {
+      const i = colors.indexOf(d.color);
+      if (i > -1) {
+        colors.splice(i, 1);
+      }
+    });
+    return colors;
+  }
+
+  protected loadSelectionColumnData(id: number): Promise<IScoreRow<any>[]> {
+    // hook
+    return Promise.resolve([]);
+  }
+
+  /**
+   *
+   * @param scoreImpl
+   * @param scorePlugin
+   */
+  protected addScoreColumn(scoreImpl: IScore<any>, scorePlugin: RangeLike) {
+    const colDesc = scoreImpl.createDesc();
+    colDesc._score = scorePlugin;
+
+    const loadScoreColumn = (id) => {
+      return scoreImpl.compute([], this.idType);
+    };
+
+    this.addColumn(colDesc, loadScoreColumn);
+  }
+
+  destroy() {
+    if (this.lineup) {
+      this.lineup.on('updateStart', null);
+      this.lineup.on('updateFinished', null);
+    }
+    super.destroy();
+  }
+
+  protected build(rows: any[] = [], columns: any[] = []) {
+    // prevent re-initialization
+    if (this.lineup) {
+      return;
+    }
+
+    deriveColors(columns);
+
+    const storage = new LocalDataProvider(rows, columns);
+    this.lineup = new LineUp(this.node, storage, this.config);
+    this.initSelectionHelper();
+
+    //this.lineup.on('updateStart', () => { this.setBusy(true); });
+    //this.lineup.on('updateFinished', () => { this.setBusy(false); });
+
+    const ranking = this.lineup.data.pushRanking();
+
+    columns.forEach((d, i) => {
+      // add visible columns
+      if (d.visible) {
+        this.lineup.data.push(ranking, d);
+      }
+
+      // set initial column width
+      if (d.width > -1) {
+        ranking.children[i + 1].setWidth(d.width); // i+1 because first column == rank
+      }
+    });
+
+    // add selection column
+    useDefaultLayout(this.lineup);
+    this.lineup.update();
+  }
+
+  protected update() {
+    this.setBusy(true);
+
+    this.initLineUpPromise = this.loadColumnDesc()
+      .then((desc: {idType: string, columns: any[]}) => {
+        this.initColumns(desc);
+        return this.loadRows();
+      })
+      .then((rows: any[]) => {
+        this.initRows(rows);
+      })
+      .then(() => {
+        this.initializedLineUp();
+        this.setBusy(false);
+      })
+      .catch(() => {
+        this.setBusy(false);
+      });
+  }
+
+  protected loadColumnDesc() {
+    // hook
+    return new Promise((resolve) => {
+      const r = {
+        idType: '',
+        columns: []
+      };
+      resolve(r);
+    });
+  }
+
+  protected initColumns(desc: { idType: string}) {
+    this.idType = resolve(desc.idType);
+  }
+
+  protected loadRows() {
+    return new Promise((resolve, reject) => {
+      const r = [];
+      resolve(r);
+    });
+  }
+
+  protected initRows(rows: any[]) {
+    this.$nodata.classed('hidden', (rows.length > 0));
+    // no data available
+    if (rows.length === 0) {
+      return [];
+    }
+
+    rows = this.mapRows(rows);
+
+    this.fillIDTypeMapCache(this.idType, rows);
+    const provider = <LocalDataProvider>this.lineup.data;
+    provider.setData(rows);
+    this.selectionHelper.rows = rows;
+    return rows;
+  }
+
+  protected mapRows(rows: any[]) {
+    // hook
+    return rows;
+  }
+
+  private updateRef(storage) {
+    if (this.resolver) {
+      this.resolver(storage);
+      this.resolver = null;
+    }
+    this.context.ref.value.data = Promise.resolve(storage);
+  }
+
+  protected initializedLineUp() {
+    this.updateRef(this.lineup.data);
+    cmds.clueify(this.context.ref, this.context.graph);
+    this.updateLineUpStats();
+  }
+
+  protected withoutTracking(f: (lineup: any)=>void) {
+    cmds.untrack(this.context.ref)
+      .then(f.bind(this, this.lineup))
+      .then(cmds.clueify.bind(cmds, this.context.ref, this.context.graph));
+  }
+
+  /**
+   * Fill the idType map cache
+   * The goal is to avoid further mapping request from the id resolver to the server
+   * @param idtype
+   * @param rows
+   */
+  protected fillIDTypeMapCache(idtype: IDType, rows: {_id: number, id: string}[]) {
+    const ids = [], names = [];
+    rows.forEach((r, i) => {
+      ids[i] = r._id;
+      names[i] = r.id;
+    });
+    idtype.fillMapCache(ids, names);
+  }
+
+  /**
+   * Writes the number of total, selected and shown items in the parameter area
+   */
+  updateLineUpStats() {
+
+    /**
+     * Builds the stats string
+     * @param total
+     * @param selected
+     * @param shown
+     * @returns {string}
+     */
+    const showStats = (total: number, selected = 0, shown = 0) => {
+      let str = 'Showing ';
+
+      str += `${shown} `;
+      if (total !== 0) {
+        str += `of ${total} `;
+      }
+      str += this.getItemName(total);
+      if (selected > 0) {
+        str += `; ${selected} selected`;
+      }
+      return str;
+    };
+
+
+    // this.lineup not available
+    if (!this.lineup) {
+      this.$params.html(showStats(0, 0));
+      return;
+    }
+
+    const selected = this.lineup.data.getSelection().length;
+    const total = (<LocalDataProvider>this.lineup.data).data.length;
+
+    const r = this.lineup.data.getRankings()[0];
+    if (r) {
+      // needs a setTimeout, because LineUp needs time to filter the rows
+      const id = setTimeout(() => {
+        clearTimeout(id);
+        const shown = r.getOrder().length;
+        this.$params.html(showStats(total, selected, shown));
+      }, 150); // adjust the time depending on the number of rows(?)
+    } else {
+      this.$params.html(showStats(total, selected));
+    }
+  }
+
+  protected getItemName(count: number): string {
+    return (count === 1) ? 'item' : 'items';
+  }
+
+  /**
+   * Destroy LineUp instance
+   */
+  protected clear() {
+    if (this.lineup) {
+      this.lineup.destroy();
+      this.lineup = undefined; // delete ref to call this.build() again
+
+      this.selectionHelper.destroy();
+      this.selectionHelper = undefined;
+    }
+  }
+
+}
+
+export default ALineUpView2;
